requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
chromadb
ollama
python-dotenv
sentence-transformers
pypdf